+++ client/src/api/users.js
@@ -32,4 +32,23 @@ export default {
   acceptTermsOfService: () => {
     return post('/api/users/me/terms_of_services', {})
   },
+
+  resetPassword: (email) => {
+    const payload = {
+      user: {
+        email,
+      },
+    }
+    return post('/api/users/password_resets', payload, false)
+  },
+
+  changePassword: (token, password) => {
+    const payload = {
+      user: {
+        password,
+      },
+      token,
+    }
+    return post('/api/users/passwords', payload, false)
+  },
 }